Now compatible with MicroPython
I have made some changes to make this library compatible with micropython.
The pca9685 library was available for circuit python only earlier but after some tweaking I could run it on micropython. After importing the library into your code and making a code using this library, the code will instantly look for a I2C device, so you might want to connect the board and then run the test as otherwise you might get an error.